desktop/requirements.txt
pystray==0.19.5
Pillow==10.4.0
psutil==5.9.8
requests==2.32.3
google-generativeai==0.7.2
uiautomation==2.0.18
